Project Start
A simple board class which just holds everything together. A main
(picross) program that takes in a file and prints it and a test file. A
file should be in the form x size, y size, x clues with each line being
a separate column, y clues with each line being a separate row. Next, I
will start to implement human input which will lead to the program
solving it itself.